+++ src/test/java/gdsc/binaryho/imhere/core/auth/AuthServiceTest.java
@@ -1,7 +1,6 @@
 package gdsc.binaryho.imhere.core.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -9,7 +8,6 @@ import static org.junit.jupiter.api.Assertions.fail;
 import gdsc.binaryho.imhere.core.auth.application.AuthService;
 import gdsc.binaryho.imhere.core.auth.application.port.VerificationCodeRepository;
 import gdsc.binaryho.imhere.core.auth.exception.DuplicateEmailException;
-import gdsc.binaryho.imhere.core.auth.exception.EmailVerificationCodeIncorrectException;
 import gdsc.binaryho.imhere.core.auth.exception.MemberNotFoundException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordFormatMismatchException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordIncorrectException;
@@ -42,7 +40,6 @@ class AuthServiceTest {
     private static final String NAME = "이진호";
     private static final String PASSWORD = "abcd1234";
     private static final String DEFAULT_MEMBER_ROLE = "ROLE_STUDENT";
-    private static final String EMAIL = "dlwlsgh4687@gmail.com";
 
     @BeforeEach
     void initAuthService() {
@@ -118,31 +115,5 @@ class AuthServiceTest {
             authService.validateSignInRequest(new SignInRequest(UNIV_ID, PASSWORD));
 
         assertThat(signInRequestValidationResult.getRoleKey()).isEqualTo(DEFAULT_MEMBER_ROLE);
-    }
-
-    @Test
-    void 인증_코드를_인증할_수_있다() {
-        // given
-        String verificationCode = "imhere forever";
-        verificationCodeRepository.saveWithEmailAsKey(EMAIL, verificationCode);
-
-        // when
-        // then
-        assertThatCode(
-            () -> authService.verifyCode(EMAIL, verificationCode)
-        ).doesNotThrowAnyException();
-    }
-
-    @Test
-    void 인증_코드가_틀린_경우_예외를_발생시킨다() {
-        // given
-        String verificationCode = "imhere forever";
-        verificationCodeRepository.saveWithEmailAsKey(EMAIL, verificationCode);
-
-        // when
-        // then
-        assertThatThrownBy(
-            () -> authService.verifyCode(EMAIL, verificationCode + "wrong code")
-        ).isInstanceOf(EmailVerificationCodeIncorrectException.class);
     }
 }
